make all lib files usable without an import_map
