+++ import_map.json
@@ -23,16 +23,13 @@
 
     "@js-temporal/polyfill": "https://esm.sh/@js-temporal/polyfill@0.4.4",
     "cliffy/": "https://deno.land/x/cliffy@v1.0.0-rc.3/",
-    "croner": "https://deno.land/x/croner@4.4.1/src/croner.js",
     "denosass": "https://deno.land/x/denosass@1.0.4/mod.ts",
     "djwt": "https://deno.land/x/djwt@v3.0.1/mod.ts",
-    "dynamodb": "https://denopkg.com/chiefbiiko/dynamodb@55e60a5/mod.ts",
     "fake-indexeddb/": "https://esm.sh/fake-indexeddb@4.0.2/",
     "fast-diff": "https://esm.sh/fast-diff@1.3.0",
     "fuse": "https://deno.land/x/fuse@v6.4.1/dist/fuse.esm.min.js",
     "gitignore-parser": "https://esm.sh/gitignore-parser@0.0.2",
     "hono/": "https://deno.land/x/hono@v3.12.2/",
-    "idb/": "https://esm.sh/idb@7.1.1/",
     "js-yaml": "https://esm.sh/js-yaml@4.1.0",
     "mimetypes": "https://deno.land/x/mimetypes@v1.0.0/mod.ts",
     "porter-stemmer": "https://esm.sh/porter-stemmer@0.9.1",
