+++ lib/data/datastore.ts
@@ -1,8 +1,8 @@
-import { applyQueryNoFilterKV } from "$sb/lib/query.ts";
+import { applyQueryNoFilterKV } from "../../plug-api/lib/query.ts";
 import { FunctionMap, KV, KvKey, KvQuery } from "../../plug-api/types.ts";
-import { builtinFunctions } from "$lib/builtin_query_functions.ts";
+import { builtinFunctions } from "../builtin_query_functions.ts";
 import { KvPrimitives } from "./kv_primitives.ts";
-import { evalQueryExpression } from "$sb/lib/query_expression.ts";
+import { evalQueryExpression } from "../../plug-api/lib/query_expression.ts";
 /**
  * This is the data store class you'll actually want to use, wrapping the primitives
  * in a more user-friendly way
