+++ lib/deps.ts
@@ -6,3 +6,12 @@ export { mime } from "https://deno.land/x/mimetypes@v1.0.0/mod.ts";
 export * as esbuild from "https://deno.land/x/esbuild@v0.20.0/mod.js";
 export { denoPlugins } from "https://deno.land/x/esbuild_deno_loader@0.8.5/mod.ts";
 export * as YAML from "https://deno.land/std@0.184.0/yaml/mod.ts";
+export {
+  createClient,
+  type DynamoDBClient,
+} from "https://denopkg.com/chiefbiiko/dynamodb@55e60a5/mod.ts";
+export {
+  type IDBPDatabase,
+  openDB,
+} from "https://esm.sh/idb@7.1.1/with-async-ittr";
+export { Cron } from "https://deno.land/x/croner@4.4.1/src/croner.js";
