+++ lib/plugos/namespace.ts
@@ -1,0 +1,6 @@
+export type NamespaceOperation =
+  | "readFile"
+  | "writeFile"
+  | "listFiles"
+  | "getFileMeta"
+  | "deleteFile";
