+++ lib/plugos/syscalls/fetch.ts
@@ -1,5 +1,5 @@
 import type { SysCallMapping } from "../system.ts";
-import { ProxyFetchRequest, ProxyFetchResponse } from "$common/proxy_fetch.ts";
+import { ProxyFetchRequest, ProxyFetchResponse } from "../../proxy_fetch.ts";
 import { base64Encode } from "../../crypto.ts";
 
 export function sandboxFetchSyscalls(): SysCallMapping {
