+++ lib/proxy_fetch.ts
@@ -1,4 +1,4 @@
-import { base64Decode, base64Encode } from "$lib/crypto.ts";
+import { base64Decode, base64Encode } from "./crypto.ts";
 
 export type ProxyFetchRequest = {
   method?: string;
